followers_clear/requirements.txt
python-twitter
requests
requests_oauthlib
pyyaml